chore: cleanup rest server
