+++ delta-sharing/core/Cargo.toml
@@ -7,7 +7,6 @@ authors = ["Robert Pack <robstar.pack@gmail.com>"]
 [dependencies]
 delta-sharing-common = { path = "../common" }
 delta-sharing-client = { path = "../client" }
-delta-sharing-server = { path = "../server" }
 
 # workspace dependencies (in alphabetical order)
 chrono.workspace = true
@@ -17,4 +16,4 @@ serde_json.workspace = true
 thiserror.workspace = true
 
 tokio = { version = "1", features = ["full"] }
-serde_yml = { version = "0.0.11" }
+serde_yml = { version = "0.0.12" }
