--- delta-sharing/server/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "delta-sharing-server"
-version = "0.1.0"
-edition = "2021"
-authors = ["Robert Pack <robstar.pack@gmail.com>"]
-
-[dependencies]
-delta-sharing-common = { path = "../common" }
-
-# workspace dependencies (in alphabetical order)
-async-trait = { workspace = true }
-clap = { workspace = true }
-futures = { workspace = true }
-hyper = { workspace = true }
-delta_kernel = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tonic = { workspace = true }
-tower = { workspace = true }
-tracing = { workspace = true }
-
-# server dependencies (in alphabetical order)
-axum = "0.7.5"
-futures-util = "0.3.28"
-pin-project = "1.1"
-serde_yml = { version = "0.0.12" }
-tokio = { version = "1.10.0", features = ["full"] }
-tower-http = { version = "0.6.2", features = ["trace"] }
-http-body-util = { version = "0.1" }
-http-body = { version = "1" }
-
-[dev-dependencies]
-tower = "*"
-http = "*"
-http-body-util = "*"
-serde_json = "*"
